Added a Static Room Visibility component for directly placed actors in room levels that don't get out
