+++ Source/ProceduralDungeon/Public/RoomVisibilityComponent.h
@@ -25,66 +25,20 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "Components/ActorComponent.h"
+#include "Components/StaticRoomVisibilityComponent.h"
 #include "RoomVisitor.h"
 #include "RoomVisibilityComponent.generated.h"
 
-UENUM(BlueprintType, meta = (DisplayName = "Room Visibility"))
-enum class EVisibilityMode : uint8
-{
-	Default			UMETA(DisplayName = "Same As Room"),
-	ForceVisible	UMETA(DisplayName = "Force Visible"),
-	ForceHidden		UMETA(DisplayName = "Force Hidden"),
-	Custom			UMETA(DisplayName = "Custom"),
-	NbMode			UMETA(Hidden)
-};
-
-DECLARE_DYNAMIC_MULTICAST_DELEGATE_TwoParams(FRoomVisibilityEvent, AActor*, Actor, bool, IsInVisibleRoom);
-
 UCLASS(ClassGroup = "ProceduralDungeon", meta = (BlueprintSpawnableComponent, DisplayName = "Room Visibility"))
-class PROCEDURALDUNGEON_API URoomVisibilityComponent : public UActorComponent, public IRoomVisitor
+class PROCEDURALDUNGEON_API URoomVisibilityComponent : public UStaticRoomVisibilityComponent, public IRoomVisitor
 {
 	GENERATED_BODY()
 
 public:
 	URoomVisibilityComponent();
 
-	virtual void BeginPlay() override;
-
 	//~ BEGIN IRoomVisitor
 	virtual void OnRoomEnter_Implementation(ARoomLevel* RoomLevel) override;
 	virtual void OnRoomExit_Implementation(ARoomLevel* RoomLevel) override;
 	//~ END IRoomVisitor
-
-	void SetVisible(UObject* Owner, bool Visible);
-	void ResetVisible(UObject* Owner); // Same as SetVisible(Owner, false)
-
-	// Returns true if the actor is in a visible room.
-	// Always returns true when "Occlude Dynamic Actors" is unchecked in the plugin's settings
-	// Useful with "Custom" visibility.
-	UFUNCTION(BlueprintPure, Category = "Procedural Dungeon", meta = (CompactNodeTitle = "Is In Visible Room", DisplayName = "Is In Visible Room"))
-	bool IsVisible();
-
-	UFUNCTION(BlueprintCallable, Category = "Procedural Dungeon", meta = (DisplayName = "Set Visibility"))
-	void SetVisibilityMode(EVisibilityMode Mode);
-
-	UFUNCTION(BlueprintPure, Category = "Procedural Dungeon", meta = (CompactNodeTitle = "Visibility", DisplayName = "Get Visibility"))
-	FORCEINLINE EVisibilityMode GetVisibilityMode() { return VisibilityMode; }
-
-	// Called when the visibility from rooms changed (either by a room visibility change or by this actor moving between rooms).
-	// Useful to update a "Custom" visibility.
-	UPROPERTY(BlueprintAssignable, Category = "Procedural Dungeon")
-	FRoomVisibilityEvent OnRoomVisibilityChanged;
-
-protected:
-	void UpdateVisibility();
-
-	UFUNCTION()
-	void RoomVisibilityChanged(class ARoomLevel* RoomLevel, bool IsVisible);
-
-private:
-	void CleanEnablers();
-
-	TSet<TWeakObjectPtr<UObject>> VisibilityEnablers;
-	EVisibilityMode VisibilityMode;
 };
